Project_Airbnb2/Scripts/requirements.txt
altair==4.2.2
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
atomicwrites==1.4.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
black==23.3.0
bleach==6.0.0
blinker==1.5
blis==0.7.9
boltons @ file:///croot/boltons_1677628695607/work
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175566073/work
brotlipy==0.7.0
cachetools==5.3.0
catalogue==2.0.8
certifi @ file:///croot/certifi_1671487775120/work/certifi
cffi @ file:///croot/cffi_1670423216006/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
comm==0.1.3
conda==23.3.0
conda-content-trust @ file:///tmp/abs_d9e8d879-c7bd-425e-87af-59f595eb114bfmj_f2_x/croots/recipe/conda-content-trust_1658126375626/work
conda-package-handling @ file:///croot/conda-package-handling_1666940379191/work
confection==0.0.4
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1673298758482/work
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.10.6
fonttools==4.39.2
fqdn==1.5.1
fuzzywuzzy==0.18.0
gdown==4.7.1
gdrive==0.1.5
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.43.0
google-auth==2.16.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.59.0
graphviz==0.20.1
greenlet==2.0.2
haversine==2.8.0
httplib2==0.22.0
huggingface-hub==0.13.3
idna @ file:///croot/idna_1666125574438/work
imageio==2.27.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.0.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
lightgbm==3.3.5
llvmlite==0.39.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numba @ file:///croot/numba_1670258334024/work
numexpr @ file:///croot/numexpr_1668713892693/work
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
overpy==0.6
packaging @ file:///croot/packaging_1678965316161/work
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathlib-mate==1.2.1
pathspec==0.11.1
pathy==0.10.1
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
pluggy @ file:///opt/conda/conda-bld/pluggy_1648109264631/work
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
preshed==3.0.8
prettytable==3.6.0
prometheus-client==0.16.0
prompt-toolkit==2.0.10
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycallgraph2==1.1.3
pycosat @ file:///croot/pycosat_1666805506522/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.7
pydeck==0.8.0
pydeps==1.11.2
Pygments==2.14.0
Pympler==1.0.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///root/miniconda3/conda-bld/pysocks_1614704957454/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.7
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.1
regex==2023.3.23
requests @ file:///croot/requests_1678709725998/work
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.3
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666307131939/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302244777/work
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.0
seaborn==0.12.2
SecretStorage==3.3.3
semver==2.13.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.17.0
shap @ file:///croot/shap_1668715240782/work
shapely==2.0.1
simplegeneric==0.8.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
slicer @ file:///tmp/build/80754af9/slicer_1633422823758/work
smart-open==6.3.0
smmap==5.0.0
snakeviz==2.1.1
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.47
sqlalchemy-mate==1.4.28.4
srsly==2.4.6
stack-data==0.6.2
statsmodels==0.13.5
stdlib-list==0.8.0
streamlit==1.20.0
sympy==1.11.1
terminado==0.17.1
thinc==8.1.9
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.3.21
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz @ file:///croot/toolz_1667464081129/work
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm @ file:///croot/tqdm_1679561868826/work
traitlets==5.9.0
transformers==4.27.3
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
ultralytics==8.0.58
uri-template==1.2.0
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1673575511138/work
uszipcode==1.0.1
validators==0.20.0
versioneer==0.28
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
xgboost==1.7.4
zipp==3.15.0
